Multiple Updates to the Editor and Test (Index) Vue Files and the php search file, Search functionality begins here
test.php [index.php] and edit index page both had their multiselectors updated to support the tag selectors with objects that have the name of the tag and the id it is associated with in the db.

The search functionality is started here and implements the request receiving in 'search.php'. The query now needs to be constructed and sent and the response sent back in json.

the response is sent through wth 'search.js' through ajax. Once webpack compiles the 'test.php' file's javascript, this ajax function needs to be changed to access 'ajax-request' node package version.

'fetchOptions.php' has changes that send dummy data to the site but the format is the correct json format and all future responses accessing the database should be sent back using this format.